requirements.txt
autopep8==1.6.0
black==22.1.0
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
sqlparse==0.4.2
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
yoyo-migrations==7.3.2
